tools/facebook/requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
playwright-stealth>=1.0.0
